+++ test/support/controllers/test_resource_v2_controller.ex
@@ -13,8 +13,22 @@ defmodule TestResourceV2Controller do
 
   use Bonny.ControllerV2
 
-  step Bonny.Pluggable.SkipObservedGenerations
-  step :handle_action
+  step(:send_done)
+
+  step(Bonny.Pluggable.Finalizer,
+    id: "example.com/cleanup",
+    impl: &__MODULE__.cleanup/1,
+    add_to_resource: &__MODULE__.add_finalizers?/1
+  )
+
+  step(Bonny.Pluggable.Finalizer,
+    id: "example.com/cleanup2",
+    impl: &__MODULE__.cleanup2/1,
+    add_to_resource: &__MODULE__.add_finalizers?/1
+  )
+
+  step(Bonny.Pluggable.SkipObservedGenerations)
+  step(:handle_action)
 
   def handle_action(axn, _opts) do
     respond(axn.resource, axn.action)
@@ -32,8 +46,31 @@ defmodule TestResourceV2Controller do
   end
 
   def rbac_rules() do
-    [
-      to_rbac_rule({"example.com", "testresourcev2s/status", "*"})
-    ]
+    [to_rbac_rule({"example.com", "testresourcev2s/status", "*"})]
+  end
+
+  def cleanup(axn) do
+    respond(axn.resource, :cleanup)
+    {:ok, axn}
+  end
+
+  def cleanup2(axn) do
+    respond(axn.resource, :cleanup2)
+    {:ok, axn}
+  end
+
+  def add_finalizers?(axn) do
+    axn.resource["metadata"]["annotations"]["add-finalizers"] == "True"
+  end
+
+  def send_done(axn, _) do
+    Bonny.Axn.register_after_processed(axn, fn %Bonny.Axn{resource: resource} = axn ->
+      pid = resource |> get_in(["spec", "pid"]) |> Bonny.Test.ResourceHelper.string_to_pid()
+      ref = resource |> get_in(["spec", "ref"]) |> Bonny.Test.ResourceHelper.string_to_ref()
+      name = resource |> get_in(["metadata", "name"])
+
+      send(pid, {ref, :done, name})
+      axn
+    end)
   end
 end